@@ -13,31 +13,27 @@
   "author": "Matthew Morrison",
   "license": "ISC",
   "dependencies": {
+    "axios": "^0.15.3",
     "bcrypt": "^0.8.7",
     "body-parser": "^1.15.2",
     "express": "^4.14.0",
-    "jquery": "^3.1.1",
+    "flux": "^3.1.2",
     "jsonwebtoken": "^7.2.1",
     "jwt-simple": "^0.5.1",
     "mongoose": "^4.6.8",
-    "passport": "^0.3.2",
-    "passport-jwt": "^2.2.1",
-    "passport-local": "^1.0.0"
+    "react": "^15.4.2",
+    "react-dom": "^15.4.2",
+    "react-router": "^3.0.2"
   },
   "devDependencies": {
     "babel-core": "^6.17.0",
     "babel-loader": "^6.2.5",
     "babel-preset-es2015": "^6.16.0",
     "babel-preset-react": "^6.16.0",
-    "classnames": "^2.2.5",
     "css-loader": "^0.25.0",
     "extract-text-webpack-plugin": "^1.0.1",
-    "flux": "^3.1.0",
     "node-sass": "^3.10.1",
-    "react": "^15.3.2",
-    "react-dom": "^15.3.2",
     "react-hot-loader": "^3.0.0-beta.6",
-    "react-router": "^3.0.0",
     "sass-loader": "^4.0.2",
     "style-loader": "^0.13.1",
     "webpack": "^1.13.2",
